Separate sinks between Linux and Windows
